requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.9.24
cloudinary==1.30.0
Django==4.1.3
django-shopping-cart==0.1
gunicorn==20.1.0
packaging==21.3
Pillow==9.3.0
psycopg2-binary==2.9.5
pyparsing==3.0.9
python-dotenv==0.21.0
redis==4.3.5
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
